dataset_definitions json, sorted 0*.py into functions
Added functions to reduce the length of the 0_download_dataset.py file and make it a bit easier to follow. Moved the hard-coded values out of the script into dataset_definitions.py, where the user can specify which URLs to fetch datasets from; the anomalous values within those datasets and the row indices for training and test subsets.